move Adjustment type out of mint module, since its a helper only
